requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
future==0.17.1
idna==2.8
ipykernel==5.1.1
ipython==7.6.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.14.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
junos-eznc==2.2.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
lxml==4.3.4
MarkupSafe==1.1.1
mistune==0.8.4
mypy-extensions==0.4.1
napalm==2.4.0
nbconvert==5.5.0
nbformat==4.4.0
ncclient==0.6.6
netaddr==0.7.19
netmiko==2.3.3
nornir==2.2.0
notebook==5.7.8
ntc-rosetta==0.1.0
nxapi-plumbing==0.5.2
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
pydantic==0.18.2
pyeapi==0.8.2
Pygments==2.4.2
pyIOSXR==0.53
PyNaCl==1.3.0
pyrsistent==0.15.2
pyserial==3.4
python-dateutil==2.8.0
PyXB==1.2.6
PyYAML==5.1.1
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
ruamel.yaml==0.15.97
scp==0.13.2
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
textfsm==0.4.1
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
yangify==0.1.1
yangson==1.3.40